fixed splitting of functions from main
