+++ src/delay.h
@@ -1,0 +1,12 @@
+#ifndef DELAY_H
+#define DELAY_H
+
+#include <MKL25Z4.h>
+
+// Troubleshooting
+#include <stdbool.h>
+#include <stdio.h>
+
+void delay_us(uint32_t d);
+
+#endif
